scraping/requirements.txt
# Bibliothèques HTTP et scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.6

# Base de données
pymongo==4.6.0

# Traitement de données
pandas==2.1.1

# Utilitaires
python-dotenv==1.0.0

# Dépendances supplémentaires
urllib3==2.0.7
chardet==5.2.0
idna==3.4
certifi==2023.7.22 